refactor: AI 코드 리뷰 피드백 반영
- seed.ts 경로 개선: 데이터 파일을 api 프로젝트 내부로 이동
- AdminPortfolioPage/AdminResumePage 중복 코드를 JsonEditorPage로 추출
- JsonEditorPage에 에러 로깅 추가
- portfolio.json에 startDate 필드 추가 (정렬용)
- portfolio/resume API 레거시 별칭에 @deprecated 주석 추가
