+++ apps/web/src/components/admin/index.ts
@@ -1,0 +1,1 @@
+export { JsonEditorPage, NextImage } from './JsonEditorPage';
